Pruebas Unitarias realizadas
Se implementaron nuevas clases y se realizaron las pruebas unitarias

+++ Reportes/src/main/java/com/SaboresGlobales/reportes/Reportes/DTO/ReportesRequestDTO.java
@@ -1,7 +1,5 @@
 package com.SaboresGlobales.reportes.Reportes.DTO;
 
-import java.math.BigDecimal;
-
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -13,7 +11,7 @@ import lombok.NoArgsConstructor;
 public class ReportesRequestDTO {
     @NotBlank(message = "NO PUEDE ESTAR VACIO")
     private String pedido;
-    private BigDecimal pagos;
+    private double pagos;
     private Integer inventario;
     private String delivery;
