+++ Reportes/src/main/java/com/SaboresGlobales/reportes/Reportes/DTO/ReportesResponseDTO.java
@@ -1,7 +1,5 @@
 package com.SaboresGlobales.reportes.Reportes.DTO;
 
-import java.math.BigDecimal;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -14,7 +12,7 @@ public class ReportesResponseDTO {
 
     private Long id;
     private String pedido;
-    private BigDecimal pagos;
+    private double pagos;
     private Integer inventario;
     private String delivery;
 }
